@@ -1,52 +1,55 @@
-[tool.poetry]
+[project]
 name = "modfs"
 version = "0.1.0"
 description = ""
-authors = ["J. Marcè i Igual <j.marce.i.igual@tue.nl>"]
+authors = [
+    {name = "Joan Marcè i Igual", email = "j.marce.i.igual@tue.nl"}
+]
+readme = "README.md"
+requires-python = ">=3.12"
+dynamic = ["dependencies"]
 
-[tool.poetry.dependencies]
-python = "^3.10"
-tqdm = "^4.65.0"
-pandas = "^2.2.0"
-plotly = "^5.18.0"
-kaleido = "0.2.1"
-cbor2 = "^5.4.2.post1"
-slurmer = "^1.0.7"
-networkx = "^2.8.1"
-matplotlib = "^3.5.2"
-commentjson = "^0.9.0"
-pydot = "^1.4.2"
-pygraphviz = "^1.10"
-pyarrow = "^14.0.0"
-seaborn = "^0.13.1"
-
-[tool.poetry.scripts]
+[project.scripts]
 modfs = 'modfs:main'
 
+[tool.poetry.dependencies]
+python = "~3.12"
+tqdm = "^4.67.1"
+pandas = "^2.2.3"
+plotly = "^6.0.0"
+kaleido = "0.2.1"
+cbor2 = "^5.6.5"
+slurmer = "^1.0.7"
+matplotlib = "^3.10.0"
+commentjson = "^0.9.0"
+pyarrow = "^19.0.0"
+seaborn = "^0.13.2"
+
 [tool.poetry.group.dev.dependencies]
-flakeheaven = "^3.2.1"
-pytest = "^7.0.1"
-ipython = "^8.0.1"
-black = "^22.1.0"
-Sphinx = "^4.4.0"
-sphinx-rtd-theme = "^1.0.0"
-jupyter = "^1.0.0"
+flakeheaven = "^3.3.0"
+pytest = "^8.3.4"
+ipython = "^8.32.0"
+black = "^25.1.0"
+Sphinx = "^8.1.3"
+sphinx-rtd-theme = "^3.0.2"
+jupyter = "^1.1.1"
 flake8 = "^4.0.1"
 flake8-bandit = "^2.1.2"
-flake8-builtins = "^1.5.3"
+flake8-builtins = "^2.5.0"
 flake8-bugbear = "^22.1.11"
-flake8-comprehensions = "^3.8.0"
-flake8-docstrings = "^1.6.0"
+flake8-comprehensions = "^3.16.0"
+flake8-docstrings = "^1.7.0"
 flake8-eradicate = "^1.2.0"
-flake8-isort = "^4.1.1"
+flake8-isort = "^6.1.2"
 flake8-mutable = "^1.2.0"
-flake8-pytest-style = "^1.6.0"
-flake8-deprecated = "^1.3"
+flake8-pytest-style = "^2.1.0"
+flake8-deprecated = "^2.2.1"
 darglint = "^1.8.1"
 pep8-naming = "^0.12.1"
-pylint = "^2.12.2"
-snakeviz = "^2.1.1"
-Pympler = "^1.0.1"
+pylint = "^3.3.4"
+snakeviz = "^2.2.2"
+Pympler = "^1.1"
+pep8 = "^1.7.1"
 
 [tool.flakeheaven]
 exclude = ['README.md']
