@@ -1,28 +1,17 @@
-#include "pch/containers.hpp" // Precompiled headers always go first
+#include "fms/pch/containers.hpp" // Precompiled headers always go first
 
-#include "solvers/modular_args.hpp"
+#include "fms/solvers/modular_args.hpp"
 
-ModularArgs ModularArgs::fromArgs(const commandLineArgs &args) {
-    ModularArgs result{.timer = FMS::StaticTimer(args.timeOut)};
+namespace fms::cli {
 
-    for (const auto &arg : args.modularAlgorithmOption) {
-        const auto it = std::ranges::find(arg, '=');
-        const std::string key(arg.begin(), it);
-
-        std::string value;
-        if (std::distance(it, arg.end()) > 1) {
-            value = std::string(it + 1, arg.end());
-        }
-
-        if (key == kOptStoreBounds) {
-            result.storeBounds = true;
-        } else if (key == kOptStoreSequence) {
-            result.storeSequence = true;
-        } else if (key == kOptNoSelfBounds) {
-            result.selfBounds = false;
-        } else {
-            throw std::invalid_argument(fmt::format(FMT_COMPILE("Unknown modular option {}"), key));
-        }
-    }
-    return result;
+ModularArgs ModularArgs::fromArgs(const cli::CLIArgs &args) {
+    return {
+            .storeBounds = args.modularOptions.storeBounds,
+            .storeSequence = args.modularOptions.storeSequence,
+            .selfBounds = !args.modularOptions.noSelfBounds,
+            .timer = utils::time::StaticTimer(args.modularOptions.timeOut),
+            .maxIterations = args.modularOptions.maxIterations,
+    };
 }
+
+} // namespace fms::cli
