@@ -1,9 +1,30 @@
-#include "pch/containers.hpp"
+#include "fms/pch/containers.hpp"
+#include "fms/pch/fmt.hpp"
 
-#include "utils/logger.h"
+#include "fms/utils/logger.hpp"
 
-#include <fmt/compile.h>
-#include <fmt/ostream.h>
+namespace {
+std::string_view loggingColor(const fms::utils::LOGGER_LEVEL &l) {
+    switch (l) {
+    case fms::utils::LOGGER_LEVEL::CRITICAL:
+    case fms::utils::LOGGER_LEVEL::ERROR:
+        return "\033[31m";
+    case fms::utils::LOGGER_LEVEL::WARNING:
+        return "\033[33m";
+    case fms::utils::LOGGER_LEVEL::INFO:
+        return "\033[32m";
+    case fms::utils::LOGGER_LEVEL::DEBUG:
+        return "\033[34m";
+    case fms::utils::LOGGER_LEVEL::TRACE:
+        return "\033[90m";
+    default:
+        break;
+    }
+    return "";
+}
+} // namespace
+
+namespace fms::utils {
 
 LOGGER_LEVEL &increaseVerbosity(LOGGER_LEVEL &l) {
     switch (l) {
@@ -23,13 +44,9 @@ LOGGER_LEVEL &increaseVerbosity(LOGGER_LEVEL &l) {
     return l;
 }
 
-void Logger::setVerbosity(LOGGER_LEVEL l) {
-    getInstance().level = l;
-}
+void Logger::setVerbosity(LOGGER_LEVEL l) { getInstance().level = l; }
 
-LOGGER_LEVEL Logger::getVerbosity() {
-    return getInstance().level;
-}
+LOGGER_LEVEL Logger::getVerbosity() { return getInstance().level; }
 
 Logger &Logger::getInstance() {
     static Logger instance;
@@ -38,6 +55,12 @@ Logger &Logger::getInstance() {
 
 void Logger::log(LOGGER_LEVEL l, std::string_view msg) {
     if (l <= level) {
-        fmt::print(FMT_COMPILE("[{}]: {}\n"), l, msg);
+        logWithColor(l, msg);
     }
 }
+
+void Logger::logWithColor(LOGGER_LEVEL l, std::string_view msg) {
+    fmt::print(FMT_COMPILE("{}[{}]: {}\033[m\n"), loggingColor(l), l, msg);
+}
+
+} // namespace fms::utils
